instrumentation code segregated into smaller chuncks
